Update to using the new compilation testing actions repositories in CI workflows
The `arduino/actions` repository that previously hosted the GitHub Actions actions was for development of them in their experimental stage. The actions have now progressed to the point where they can have dedicated repositories. All further development work will be done in the dedicated repository.

The default branch name of the new repositories is `main`, as opposed the the `master` used in the experimental repository.

The versions of the actions in the dedicated repository now using `sketches-reports` instead of `size-deltas-reports` as the default sketches reports path and workflow name, so it was necessary to change the names specified to the `actions/upload-artifact` action's inputs to reflect this change.