Use modern API of `arduino/compile-sketches` action
The API of the `arduino/compile-sketches` GitHub Actions action used to compile the library's example sketches in the
"Compile Examples" CI workflow has evolved over time. Although measures were taken to provide backwards compatibility for
the old API, it turns out that the GitHub Actions behavior of creating environment variables for arbitrary input names
relied upon for still recognizing the old inputs is not consistent across action types. It works fine for the Docker
container action type originally used by `arduino/compile-sketches`, but not for the new "Composite Run Steps" action
type we switched to.

This change resulted in sketch data deltas no longer being generated due to the workflow's use of the unsupported input
name `enable-size-deltas-report`. This resulted in failed runs of the "Report Size Deltas" due to the reports not having
the required deltas data.

The fix is simply to update to the new input name.

I went ahead and updated to the new YAML array style of the libraries input. There is still backwards compatibility for
the old API in this case due to the input name not having changed, but only the data format. However, the old data format
is deprecated so it's safest to update.