Add reference implementations (#138)
* Add reference implementations

These projects are fully featured and tested. In the next two weeks, they will be battle tested. 
They were developed as part of rocketry project for my student association for EuRoC competition.

They might not be the best reference implementations, but I wasn't able to find anything else working here on github. They might not be up to the standards, but they work.

* Update README.md