Only compile one board per architecture. (#184)
Compiling for numerous different boards for any given architecture (i.e. MKR 1000, MKR WiFi 1010, ... for arduino:samd) is a waste of time and computational ressources because this library does not make use of any board-specific code. If the CI compiles for one board, it compiles for any of them. Otherwise it would be a violation of the specific board of the Arduino API.