Revert "Bump codecov/codecov-action from 3 to 4 (#253)"
This reverts commit 49a65ab8f31beac949539d97cbc9406bc6566ab7.

The bump originated from an automated PR from Dependabot triggered by an accidental v4 tag push in the
codecov/codecov-action repository.

The tag has since been removed due to the v4 major version series of the action still being in the beta development
phase. That removal caused subsequent runs of the workflow to fail:

Error: Unable to resolve action `codecov/codecov-action@v4`, unable to find version `v4`

So the action ref must be reverted back to v3. It can be bumped back to the v4 major version ref following the
production 4.0.0 release in the action's repo.